+++ nuxt/index.js
@@ -1,16 +1,12 @@
-import { addComponent, addPlugin, createResolver, defineNuxtModule } from '@nuxt/kit';
+import { addComponent, createResolver, defineNuxtModule } from '@nuxt/kit';
 
 export default defineNuxtModule({
-	async setup() {
+	setup() {
 		const { resolve } = createResolver(import.meta.url);
-		await addComponent({
+		addComponent({
 			filePath: resolve('../vue/index.js'),
 			global: true,
 			name: 'PixelperfectTool',
-		});
-		addPlugin({
-			mode: 'client',
-			src: resolve('./runtime/pixelperfect-storage.client.js'),
 		});
 	},
 });