+++ nuxt/runtime/pixelperfect-storage.client.js
@@ -1,7 +1,0 @@
-/// <reference path="../../types/index.d.ts" />
-
-import { ensureValidPpOffsets } from '../../lib/ensure-pp-offsets.js';
-
-export default () => {
-	ensureValidPpOffsets();
-};